chore: global.d.ts 타입선언, 컴포넌트 이동
각 파일에 선언한 type들을 global.d.ts 파일로 이동하였으며, 기존 mypage의 components 폴더를 전체 components 폴더로 이동시켰습니다.
