+++ src/components/onboarding/DoneButton.tsx
@@ -1,14 +1,10 @@
 'use client';
 
 import React, { ReactNode } from 'react';
-import { useState } from 'react';
-interface DoneButtonProps {
-  children: ReactNode;
-  onClick?: () => void;
-  isClick: boolean;
-}
-const DoneButton: React.FC<DoneButtonProps> = ({children, onClick, isClick}) => {
 
+import { DoneButtonProps } from '@/types/global';
+
+const DoneButton: React.FC<DoneButtonProps> = ({ children, onClick, isClick }) => {
   return (
     <button
       className={
@@ -17,9 +13,7 @@ const DoneButton: React.FC<DoneButtonProps> = ({children, onClick, isClick}) => 
           : 'w-full bg-gray2 h-[100px] rounded-t-[32px] text-white text-h3 fixed bottom-0'
       }
       onClick={onClick}>
-      <div className="text-white text-h3 py-[25px]">
-        {children}
-      </div>
+      <div className="text-white text-h3 py-[25px]">{children}</div>
     </button>
   );
 };
